Fix number format bug on german devices
When using XamlReader.Load to create a Geometry element for the path / mask for a Stump or Elbo, the app might crash.
This only happens on german devices because the system there uses a comma instead of a point as decimal seperator, which will cause a parsing error there.
This does not happen very often but can happen when either a division ends up with a decimal number or a decimal number is specified in XAML.
I fixed this bug by just setting the language for the string conversion to english.